separate package for APIs and documentation automation
