+++ docs/README.md
@@ -1,0 +1,7 @@
+[machine-controller-manager](./docs/machine-controller-manager.md ':include')
+
+[gcp-oot](./docs/gcp-oot.md ':include')
+
+[provider-spec](./docs/provider-spec.md ':include')
+
+[gcp-oot-usage](./docs/gcp-oot-usage.md ':include')
