+++ hack/api-reference/generate-spec-doc.sh
@@ -1,0 +1,18 @@
+./gen-crd-api-reference-docs -config "providerspec-config.json" -api-dir "../../pkg/api/v1alpha1" -out-file="../../docs/docs/provider-spec.md"
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
