+++ hack/api-reference/providerspec-config.json
@@ -1,0 +1,43 @@
+{
+    "hideMemberFields": [
+        "TypeMeta"
+    ],
+    "hideTypePatterns": [
+        "ParseError$",
+        "List$"
+    ],
+    "externalPackages": [{
+            "typeMatchPrefix": "^k8s\\.io/apimachinery/pkg/apis/meta/v1\\.Duration$",
+            "docsURLTemplate": "https://godoc.org/k8s.io/apimachinery/pkg/apis/meta/v1#Duration"
+        },
+        {
+            "typeMatchPrefix": "^k8s\\.io/apimachinery/pkg/api/resource\\.Quantity$",
+            "docsURLTemplate": "https://godoc.org/k8s.io/apimachinery/pkg/api/resource#Quantity"
+        },
+        {
+            "typeMatchPrefix": "^k8s\\.io/apimachinery/pkg/runtime\\.RawExtension$",
+            "docsURLTemplate": "https://godoc.org/k8s.io/apimachinery/pkg/runtime#RawExtension"
+        },
+        {
+            "typeMatchPrefix": "^k8s\\.io/apimachinery/pkg/types\\.UID$",
+            "docsURLTemplate": "https://godoc.org/k8s.io/apimachinery/pkg/types#UID"
+        },
+        {
+            "typeMatchPrefix": "^k8s\\.io/apimachinery/pkg/util/intstr\\.IntOrString$",
+            "docsURLTemplate": "https://godoc.org/k8s.io/apimachinery/pkg/util/intstr#IntOrString"
+        },
+        {
+            "typeMatchPrefix": "^time\\.Duration$",
+            "docsURLTemplate": "https://godoc.org/time#Duration"
+        },
+        {
+            "typeMatchPrefix": "^k8s\\.io/(api|apimachinery/pkg/apis)/",
+            "docsURLTemplate": "https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.19/#{{lower .TypeIdentifier}}-{{arrIndex .PackageSegments -1}}-{{arrIndex .PackageSegments -2}}"
+        }
+    ],
+    "typeDisplayNamePrefixOverrides": {
+        "k8s.io/api/": "Kubernetes ",
+        "k8s.io/apimachinery/pkg/apis/": "Kubernetes "
+    },
+    "markdownDisabled": false
+}
