+++ hack/api-reference/template/pkg.tpl
@@ -1,0 +1,17 @@
+{{ define "packages" }}
+
+## Specification
+### ProviderSpec Schema
+
+{{ range .packages }}
+    {{ range (visibleTypes (sortedTypes .Types))}}
+        {{ template "type" .  }}
+    {{ end }}
+    <hr/>
+{{ end }}
+
+<p><em>
+    Generated with <a href="https://github.com/ahmetb/gen-crd-api-reference-docs">gen-crd-api-reference-docs</a>
+</em></p>
+
+{{ end }}
